Правки

+++ src/main/java/me/kirillirik/tree/Color.java
@@ -2,43 +2,43 @@ package me.kirillirik.tree;
 
 public final class Color {
 
-    private float r;
-    private float g;
-    private float b;
+    private int r;
+    private int g;
+    private int b;
 
-    public Color(float r, float g, float b) {
+    public Color(int r, int g, int b) {
         this.r = r;
         this.g = g;
         this.b = b;
     }
 
-    public void set(float r, float g, float b) {
+    public void set(int r, int g, int b) {
         this.r = r;
         this.g = g;
         this.b = b;
     }
 
-    public void setR(float r) {
+    public void setR(int r) {
         this.r = r;
     }
 
-    public void setG(float g) {
+    public void setG(int g) {
         this.g = g;
     }
 
-    public void setB(float b) {
+    public void setB(int b) {
         this.b = b;
     }
 
-    public float getR() {
+    public int getR() {
         return r;
     }
 
-    public float getG() {
+    public int getG() {
         return g;
     }
 
-    public float getB() {
+    public int getB() {
         return b;
     }
 }
